Removes extratools update regex tools with missing functions from extratools
